# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BizOS ERP API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupBody"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithPlan"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/onboard:
    post:
      operationId: completeOnboarding
      tags: [auth]
      summary: Complete first-time onboarding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingBody"
      responses:
        "200":
          description: Onboarding complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithPlan"

  # DASHBOARD
...
  /accounting/transactions:
    get:
      operationId: listTransactions
      tags: [accounting]
      summary: List all transactions (ledger)
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [income, expense]
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Transactions list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionsListResponse"
    post:
      operationId: createTransaction
      tags: [accounting]
      summary: Record a new transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionBody"
      responses:
        "201":
          description: Transaction recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /accounting/summary:
    get:
      operationId: getAccountingSummary
      tags: [accounting]
      summary: Get profit/loss and accounting summary
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [week, month, quarter, year]
            default: month
      responses:
        "200":
          description: Accounting summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountingSummary"

  # AI ASSISTANT
  /ai/prompt:
    post:
      operationId: sendAiPrompt
      tags: [ai]
      summary: Send a natural language prompt to the AI assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiPromptBody"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiResponse"
        "403":
          description: Prompt limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ai/history:
    get:
      operationId: getAiHistory
      tags: [ai]
      summary: Get AI prompt history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Prompt history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiPromptLog"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
...
    SignupBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        businessName:
          type: string
      required: [name, email, password, businessName]
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserWithPlan"
      required: [token, user]
    UserWithPlan:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        businessName:
          type: string
        plan:
          type: string
          enum: [free, pro]
        onboardingComplete:
          type: boolean
        aiPromptsUsed:
          type: integer
        aiPromptsLimit:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, businessName, plan, onboardingComplete, aiPromptsUsed, aiPromptsLimit, createdAt]
    OnboardingBody:
      type: object
      properties:
        businessName:
          type: string
        businessType:
          type: string
        gstNumber:
          type: string
        address:
          type: string
        phone:
          type: string
      required: [businessName, businessType]

    # DASHBOARD
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        description:
          type: string
        amount:
          type: number
        date:
          type: string
          format: date
        invoiceId:
          type: integer
        referenceNumber:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, type, category, description, amount, date, createdAt]
    TransactionsListResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalIncome:
          type: number
        totalExpenses:
          type: number
      required: [transactions, total, page, limit, totalIncome, totalExpenses]
    CreateTransactionBody:
      type: object
      properties:
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        description:
          type: string
        amount:
          type: number
          minimum: 0
        date:
          type: string
          format: date
        invoiceId:
          type: integer
        referenceNumber:
          type: string
      required: [type, category, description, amount, date]
    AccountingSummary:
      type: object
      properties:
        period:
          type: string
        totalIncome:
          type: number
        totalExpenses:
          type: number
        netProfit:
          type: number
        profitMargin:
          type: number
        topExpenseCategories:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              amount:
                type: number
            required: [category, amount]
        incomeBreakdown:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              amount:
                type: number
            required: [category, amount]
      required: [period, totalIncome, totalExpenses, netProfit, profitMargin, topExpenseCategories, incomeBreakdown]

    # AI ASSISTANT
    AiPromptBody:
      type: object
      properties:
        prompt:
          type: string
      required: [prompt]
    AiResponse:
      type: object
      properties:
        answer:
          type: string
        intent:
          type: string
        data:
          type: object
        requiresConfirmation:
          type: boolean
        confirmationMessage:
          type: string
        promptsRemaining:
          type: integer
      required: [answer, intent, requiresConfirmation, promptsRemaining]
    AiPromptLog:
      type: object
      properties:
        id:
          type: integer
        prompt:
          type: string
        intent:
          type: string
        response:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, prompt, intent, response, createdAt]
...